fix: New payout calculation
- Moved tests into own file
