+++ src/views/Predictions/components/History/HistoricalBet.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react'
 import { useWeb3React } from '@web3-react/core'
-import { ethers } from 'ethers'
 import {
   Box,
   ChevronDownIcon,
@@ -81,8 +80,6 @@ const HistoricalBet: React.FC<BetProps> = ({ bet }) => {
 
   // Winners get the payout, otherwise the claim what they put it if it was canceled
   const payout = roundResult === Result.WIN ? getNetPayout(bet, rewardRate) : amount
-  // TODO: remove this when we don't have to convert
-  const payoutAsEthBn = ethers.BigNumber.from(payout)
 
   const renderBetLabel = () => {
     if (isOpenRound) {
@@ -137,7 +134,7 @@ const HistoricalBet: React.FC<BetProps> = ({ bet }) => {
           <CollectWinningsButton
             hasClaimed={!canClaim}
             epoch={bet.round.epoch}
-            payout={payoutAsEthBn}
+            payout={payout.toString()}
             scale="sm"
             mr="8px"
           >
