+++ src/views/Predictions/components/RoundCard/ExpiredRoundCard.tsx
@@ -4,9 +4,9 @@ import { useWeb3React } from '@web3-react/core'
 import { Box, BlockIcon, CardBody } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { NodeRound, BetPosition, NodeLedger } from 'state/types'
-import { useBlock, useGetBetByEpoch, useGetRewardRate } from 'state/hooks'
+import { useBlock, useGetBetByEpoch } from 'state/hooks'
 import { RoundResult } from '../RoundResult'
-import { getHasRoundFailed, getNetPayoutv2 } from '../../helpers'
+import { getHasRoundFailed, getPayoutv2 } from '../../helpers'
 import MultiplierArrow from './MultiplierArrow'
 import Card from './Card'
 import CardHeader from './CardHeader'
@@ -46,8 +46,7 @@ const ExpiredRoundCard: React.FC<ExpiredRoundCardProps> = ({
 
   const betPosition = closePrice > lockPrice ? BetPosition.BULL : BetPosition.BEAR
   const ledger = useGetBetByEpoch(account, epoch)
-  const rewardRate = useGetRewardRate()
-  const payout = getNetPayoutv2(ledger, round, rewardRate)
+  const payout = getPayoutv2(ledger, round)
   const hasRoundFailed = getHasRoundFailed(round, initialBlock)
 
   if (hasRoundFailed) {
